Move breadcrumb pickers into a new breadcrumb_picker crate
picker keeps a dev-dependency on editor for its own tests, so editor
depending on picker linked two copies of picker into picker's test
binary and every test panicked on duplicate action registration.
editor now exposes a OnceLock the new crate fills at startup, the way
outline registers TOGGLE_OUTLINE.
