+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
     "crates/aws_http_client",
     "crates/bedrock",
     "crates/benchmarks",
+    "crates/breadcrumb_picker",
     "crates/breadcrumbs",
     "crates/buffer_diff",
     "crates/call",
@@ -291,6 +292,7 @@ auto_update = { path = "crates/auto_update" }
 auto_update_ui = { path = "crates/auto_update_ui" }
 aws_http_client = { path = "crates/aws_http_client" }
 bedrock = { path = "crates/bedrock" }
+breadcrumb_picker = { path = "crates/breadcrumb_picker" }
 breadcrumbs = { path = "crates/breadcrumbs" }
 buffer_diff = { path = "crates/buffer_diff" }
 call = { path = "crates/call" }
