+++ crates/breadcrumb_picker/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "breadcrumb_picker"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/breadcrumb_picker.rs"
+doctest = false
+
+[dependencies]
+editor.workspace = true
+file_icons.workspace = true
+fuzzy.workspace = true
+gpui.workspace = true
+language.workspace = true
+menu.workspace = true
+picker.workspace = true
+project.workspace = true
+settings.workspace = true
+text.workspace = true
+theme.workspace = true
+ui.workspace = true
+util.workspace = true
+workspace.workspace = true
+zed_actions.workspace = true
+
+[dev-dependencies]
+collections.workspace = true
+editor = { workspace = true, features = ["test-support"] }
+gpui = { workspace = true, features = ["test-support"] }
+language = { workspace = true, features = ["test-support"] }
+project = { workspace = true, features = ["test-support"] }
+serde_json.workspace = true
+settings = { workspace = true, features = ["test-support"] }
+theme = { workspace = true, features = ["test-support"] }
+theme_settings.workspace = true
+workspace = { workspace = true, features = ["test-support"] }
