+++ crates/editor/Cargo.toml
@@ -65,7 +65,6 @@ multi_buffer.workspace = true
 ordered-float.workspace = true
 parking_lot.workspace = true
 pretty_assertions.workspace = true
-picker.workspace = true
 project.workspace = true
 proptest = { workspace = true, optional = true }
 proptest-derive = { workspace = true, optional = true }
