+++ crates/zed/Cargo.toml
@@ -76,6 +76,7 @@ assets.workspace = true
 audio.workspace = true
 auto_update.workspace = true
 auto_update_ui.workspace = true
+breadcrumb_picker.workspace = true
 breadcrumbs.workspace = true
 call.workspace = true
 chrono.workspace = true
